fix a bunch of 64-bit sign extension warnings, reported by Sun Studio option, -xport64=implicit
git-svn-id: file:///home/asfernandes/fb/repository/sf/svn/firebird/trunk@45564 de594faa-8d1b-4a0c-9a6a-a7de5f8bf859

+++ src/dsql/dsql.cpp
@@ -2606,7 +2606,7 @@ static dsql_req* prepare(thread_db* tdbb, dsql_dbb* database, jrd_tra* transacti
 	// restore warnings (if there are any)
 	if (local_status[2] == isc_arg_warning)
 	{
-		int indx, len, warning;
+		size_t indx, len, warning;
 
 		// find end of a status vector
 		PARSE_STATUS(tdbb->tdbb_status_vector, indx, warning);
